Restructure few packages
