+++ cmd/productive-cli/main.go
@@ -2,14 +2,13 @@ package main
 
 import (
 	"fmt"
+	"github.com/mister11/productive-cli/internal/interfaces/cli"
 	"os"
-
-	"github.com/mister11/productive-cli/internal/interfaces"
 )
 
 func main() {
 
-	if err := interfaces.CLI().Run(os.Args); err != nil {
+	if err := cli.Run(os.Args); err != nil {
 		fmt.Fprint(os.Stderr, err)
 		os.Exit(-1)
 	}
