+++ internal/interfaces/cli/commands.go
@@ -1,4 +1,4 @@
-package interfaces
+package cli
 
 import (
 	"github.com/urfave/cli/v2"
